almost done with shop

+++ VexusCore/src/main/java/pw/vexus/core/shop/SellableItem.java
@@ -6,5 +6,5 @@ import org.bukkit.Material;
 @Value
 class SellableItem {
     private Material material;
-    private byte dataValue;
+    private short dataValue;
 }
